Fix ICU 78 CodeQL format-string type mismatches (E1+E2)
Resolves 13 cpp/wrong-type-format-argument alerts across 5 test files,
each mirroring an existing in-tree idiom:
- dtfmrgts.cpp:1728  %d -> %f  (d is UDate/double)
- cloctst.c:550/2569  int32_t casts for size_t args
- ccaltst.c:306/310/2950  austrdup() for UChar* printed with %s
- ccaltst.c:647/670  actual=%d -> %s (args already austrdup-wrapped)
- cnmdptst.c:267/736  austrdup() for UChar*
- creststn.c:2962  %S -> %s with a null-safe austrdup guard (idiom from cnumtst.c:3664)

These are long-standing upstream bugs (not part of #153). Implemented by
GPT-5.5 after an independent blind plan; verified against source.

+++ icu/icu4c/source/test/cintltst/cloctst.c
@@ -547,7 +547,7 @@ static void TestPrefixes(void) {
                 
                 if(len != (int32_t)strlen(buf)) {
                     log_err("#%d: %s on %s: -> [%s] (length returned %d, actual %d!)\n",
-                        row, testTitles[n], loc, buf, len, strlen(buf)+1);
+                        row, testTitles[n], loc, buf, len, (int32_t)strlen(buf)+1);
                     
                 }
                 
@@ -2565,8 +2565,8 @@ static void TestCanonicalizationBuffer(void)
     }
 
     if (reslen != (int32_t)len) {
-        log_err("FAIL: uloc_canonicalize(%s) => \"%i\", expected \"%u\"\n",
-                name, reslen, len);
+        log_err("FAIL: uloc_canonicalize(%s) => \"%i\", expected \"%i\"\n",
+                name, reslen, (int32_t)len);
         return;
     }
